Factor out Volume4D KML operations from make_flight_intent_kml
