+++ monitoring/monitorlib/kml.py
@@ -1,6 +1,17 @@
-from pykml import parser
 import re
+from typing import List, Optional
 
+import s2sphere
+from pykml import parser
+from pykml.factory import KML_ElementMaker as kml
+
+from monitoring.monitorlib.geo import (
+    Altitude,
+    AltitudeDatum,
+    DistanceUnits,
+    egm96_geoid_offset,
+)
+from monitoring.monitorlib.geotemporal import Volume4D
 
 KML_NAMESPACE = {"kml": "http://www.opengis.net/kml/2.2"}
 
@@ -111,3 +122,139 @@ def get_kml_content(kml_file, from_string=False):
         if folder_details:
             kml_content.update(folder_details)
     return kml_content
+
+
+def _altitude_mode_of(altitude: Altitude) -> str:
+    if altitude.reference == AltitudeDatum.W84:
+        return "absolute"
+    else:
+        raise NotImplementedError(
+            f"Altitude reference {altitude.reference} not yet supported"
+        )
+
+
+def _altitude_value_of(altitude: Altitude) -> float:
+    if altitude.units == DistanceUnits.M:
+        return altitude.value
+    else:
+        raise NotImplementedError(f"Altitude units {altitude.units} not yet supported")
+
+
+def make_placemark_from_volume(
+    v4: Volume4D,
+    name: Optional[str] = None,
+    style_url: Optional[str] = None,
+    description: Optional[str] = None,
+) -> kml.Placemark:
+    if "outline_polygon" not in v4.volume or not v4.volume.outline_polygon:
+        raise NotImplementedError("Volume footprint type not supported")
+
+    # Create placemark
+    args = []
+    if name is not None:
+        args.append(kml.name(name))
+    if style_url is not None:
+        args.append(kml.styleUrl(style_url))
+    placemark = kml.Placemark(*args)
+    if description:
+        placemark.append(kml.description(description))
+
+    # Set time range
+    timespan = None
+    if "time_start" in v4 and v4.time_start:
+        timespan = kml.TimeSpan(kml.begin(v4.time_start.datetime.isoformat()))
+    if "time_end" in v4 and v4.time_end:
+        if timespan is None:
+            timespan = kml.TimeSpan()
+        timespan.append(kml.end(v4.time_end.datetime.isoformat()))
+    if timespan is not None:
+        placemark.append(timespan)
+
+    # Create top and bottom of the volume
+    vertices = v4.volume.outline_polygon.vertices
+    avg = s2sphere.LatLng.from_degrees(
+        lat=sum(v.lat for v in vertices) / len(vertices),
+        lng=sum(v.lng for v in vertices) / len(vertices),
+    )
+    geoid_offset = egm96_geoid_offset(avg)
+    lower_coords = []
+    upper_coords = []
+    alt_lo = _altitude_value_of(v4.volume.altitude_lower) - geoid_offset
+    alt_hi = _altitude_value_of(v4.volume.altitude_upper) - geoid_offset
+    for vertex in vertices:
+        lower_coords.append((vertex.lng, vertex.lat, alt_lo))
+        upper_coords.append((vertex.lng, vertex.lat, alt_hi))
+    geo = kml.MultiGeometry(
+        kml.Polygon(
+            kml.altitudeMode(_altitude_mode_of(v4.volume.altitude_lower)),
+            kml.outerBoundaryIs(
+                kml.LinearRing(
+                    kml.coordinates(
+                        " ".join(",".join(str(v) for v in c) for c in lower_coords)
+                    )
+                )
+            ),
+        ),
+        kml.Polygon(
+            kml.altitudeMode(_altitude_mode_of(v4.volume.altitude_upper)),
+            kml.outerBoundaryIs(
+                kml.LinearRing(
+                    kml.coordinates(
+                        " ".join(",".join(str(v) for v in c) for c in upper_coords)
+                    )
+                )
+            ),
+        ),
+    )
+
+    # We can only create the sides of the volume if the altitude references are the same
+    if v4.volume.altitude_lower.reference == v4.volume.altitude_upper.reference:
+        indices = list(range(len(vertices)))
+        for i1, i2 in zip(indices, indices[1:] + [0]):
+            coords = [
+                (vertices[i1].lng, vertices[i1].lat, alt_lo),
+                (vertices[i1].lng, vertices[i1].lat, alt_hi),
+                (vertices[i2].lng, vertices[i2].lat, alt_hi),
+                (vertices[i2].lng, vertices[i2].lat, alt_lo),
+            ]
+            geo.append(
+                kml.Polygon(
+                    kml.altitudeMode(_altitude_mode_of(v4.volume.altitude_lower)),
+                    kml.outerBoundaryIs(
+                        kml.LinearRing(
+                            kml.coordinates(
+                                " ".join(",".join(str(v) for v in c) for c in coords)
+                            )
+                        )
+                    ),
+                )
+            )
+
+    placemark.append(geo)
+    return placemark
+
+
+def flight_planning_styles() -> List[kml.Style]:
+    """Provides KML styles with names in the form {FlightPlanState}_{AirspaceUsageState}."""
+    return [
+        kml.Style(
+            kml.LineStyle(kml.color("ff00c000"), kml.width(3)),
+            kml.PolyStyle(kml.color("80808080")),
+            id="Planned_Nominal",
+        ),
+        kml.Style(
+            kml.LineStyle(kml.color("ff00c000"), kml.width(3)),
+            kml.PolyStyle(kml.color("8000ff00")),
+            id="InUse_Nominal",
+        ),
+        kml.Style(
+            kml.LineStyle(kml.color("ff00ffff"), kml.width(5)),
+            kml.PolyStyle(kml.color("8000ff00")),
+            id="InUse_OffNominal",
+        ),
+        kml.Style(
+            kml.LineStyle(kml.color("ff0000ff"), kml.width(5)),
+            kml.PolyStyle(kml.color("8000ff00")),
+            id="InUse_Contingent",
+        ),
+    ]
